[#10964] test(idp-basic): Merge IdP REST operation Jersey tests
Consolidate user and group operation tests into TestIdpOperations.
